feat: add unit testing
