+++ index.html
@@ -28,6 +28,6 @@
         </div>
 
     </section>
-<script src="index.js"></script>
+<script src="app.js" type="module"></script>
 </body>
 </html>
